refactor(harness): move workspace path-gating policy out of types.rs
workspace/types.rs mixed plain type definitions with the
allows/enforce path-gating logic and its lexical path-normalization
helpers — the security-relevant behavior the module exists to
provide. Moved into policy.rs, leaving types.rs with just
WorkspaceDescriptor's fields and the WorkspaceIsolation trait. No
behavior change.
